Use an abstraction for MOJO's which start a WildFly server.
