+++ plugin/src/main/java/org/wildfly/plugin/server/RunMojo.java
@@ -23,48 +23,28 @@
 package org.wildfly.plugin.server;
 
 import java.io.File;
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import javax.inject.Inject;
 
-import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugin.logging.Log;
-import org.apache.maven.plugins.annotations.Component;
 import org.apache.maven.plugins.annotations.Execute;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
-import org.apache.maven.project.MavenProject;
-import org.eclipse.aether.RepositorySystem;
-import org.eclipse.aether.RepositorySystemSession;
-import org.eclipse.aether.repository.RemoteRepository;
 import org.jboss.as.controller.client.ModelControllerClient;
-import org.jboss.galleon.ProvisioningException;
-import org.jboss.galleon.maven.plugin.util.MavenArtifactRepositoryManager;
-import org.jboss.galleon.universe.maven.repo.MavenRepoManager;
 import org.wildfly.core.launcher.CommandBuilder;
-import org.wildfly.core.launcher.Launcher;
-import org.wildfly.core.launcher.StandaloneCommandBuilder;
 import org.wildfly.plugin.cli.CommandConfiguration;
 import org.wildfly.plugin.cli.CommandExecutor;
-import org.wildfly.plugin.common.AbstractServerConnection;
 import org.wildfly.plugin.common.PropertyNames;
-import org.wildfly.plugin.common.Utils;
 import org.wildfly.plugin.core.Deployment;
 import org.wildfly.plugin.core.DeploymentManager;
-import org.wildfly.plugin.core.GalleonUtils;
-import org.wildfly.plugin.core.MavenRepositoriesEnricher;
-import org.wildfly.plugin.core.ServerHelper;
 import org.wildfly.plugin.deployment.PackageType;
 
 /**
@@ -77,78 +57,10 @@ import org.wildfly.plugin.deployment.PackageType;
  */
 @Mojo(name = "run", requiresDependencyResolution = ResolutionScope.RUNTIME)
 @Execute(phase = LifecyclePhase.PACKAGE)
-public class RunMojo extends AbstractServerConnection {
-
-    @Component
-    RepositorySystem repoSystem;
-
-    @Parameter(defaultValue = "${repositorySystemSession}", readonly = true, required = true)
-    private RepositorySystemSession session;
-
-    @Parameter(defaultValue = "${project.remoteProjectRepositories}", readonly = true, required = true)
-    private List<RemoteRepository> repositories;
-
-    @Parameter(defaultValue = "${project}", readonly = true, required = true)
-    MavenProject project;
-
-    @Parameter(defaultValue = "${session}", readonly = true, required = true)
-    MavenSession mavenSession;
+public class RunMojo extends AbstractServerStartMojo {
 
     @Inject
     private CommandExecutor commandExecutor;
-
-    /**
-     * The WildFly Application Server's home directory. If not used, WildFly will be installed.
-     */
-    @Parameter(alias = "jboss-home", property = PropertyNames.JBOSS_HOME)
-    private String jbossHome;
-
-    /**
-     * The feature pack location. See the <a href="https://docs.wildfly.org/galleon/#_feature_pack_location">documentation</a>
-     * for details on how to format a feature pack location.
-     * <p>
-     * Note that if you define the version in the feature pack location, e.g. {@code #26.1.1.Final}, the {@code version}
-     * configuration parameter should be left blank.
-     * </p>
-     */
-    @Parameter(alias = "feature-pack-location", property = PropertyNames.WILDFLY_FEATURE_PACK_LOCATION)
-    private String featurePackLocation;
-
-    /**
-     * The version of the WildFly default server to install in case no jboss-home has been set
-     * and no server has previously been provisioned.
-     * The latest stable version is resolved if left blank.
-     */
-    @Parameter(property = PropertyNames.WILDFLY_VERSION)
-    private String version;
-
-    /**
-     * The directory name inside the buildDir where to provision the default server.
-     * By default the server is provisioned into the 'server' directory.
-     *
-     * @since 3.0
-     */
-    @Parameter(alias = "provisioning-dir", property = PropertyNames.WILDFLY_PROVISIONING_DIR, defaultValue = Utils.WILDFLY_DEFAULT_DIR)
-    private String provisioningDir;
-
-    /**
-     * The modules path or paths to use. A single path can be used or multiple paths by enclosing them in a paths
-     * element.
-     */
-    @Parameter(alias = "modules-path", property = PropertyNames.MODULES_PATH)
-    private ModulesPath modulesPath;
-
-    /**
-     * The JVM options to use.
-     */
-    @Parameter(alias = "java-opts", property = PropertyNames.JAVA_OPTS)
-    private String[] javaOpts;
-
-    /**
-     * The {@code JAVA_HOME} to use for launching the server.
-     */
-    @Parameter(alias = "java-home", property = PropertyNames.JAVA_HOME)
-    private String javaHome;
 
     /**
      * The CLI commands to execute before the deployment is deployed.
@@ -167,43 +79,6 @@ public class RunMojo extends AbstractServerConnection {
      */
     @Parameter(alias = "server-config", property = PropertyNames.SERVER_CONFIG)
     private String serverConfig;
-
-    /**
-     * The path to the system properties file to load.
-     */
-    @Parameter(alias = "properties-file", property = PropertyNames.PROPERTIES_FILE)
-    private String propertiesFile;
-
-    /**
-     * The timeout value to use when starting the server.
-     */
-    @Parameter(alias = "startup-timeout", defaultValue = "60", property = PropertyNames.STARTUP_TIMEOUT)
-    private long startupTimeout;
-
-    /**
-     * The arguments to be passed to the server.
-     */
-    @Parameter(alias = "server-args", property = PropertyNames.SERVER_ARGS)
-    private String[] serverArgs;
-
-    /**
-     * The users to add to the server.
-     */
-    @Parameter(alias = "add-user", property = "wildfly.add-user")
-    private AddUser addUser;
-
-    /**
-     * Specifies the environment variables to be passed to the process being started.
-     * <div>
-     * <pre>
-     * &lt;env&gt;
-     *     &lt;HOME&gt;/home/wildfly/&lt;/HOME&gt;
-     * &lt;/env&gt;
-     * </pre>
-     * </div>
-     */
-    @Parameter
-    private Map<String, String> env;
 
     /**
      * Specifies the name used for the deployment.
@@ -240,52 +115,28 @@ public class RunMojo extends AbstractServerConnection {
     @Parameter(property = PropertyNames.DEPLOYMENT_FILENAME)
     private String filename;
 
-    /**
-     * Set to {@code true} if you want the deployment to be skipped, otherwise {@code false}.
-     */
-    @Parameter(defaultValue = "false", property = PropertyNames.SKIP)
-    private boolean skip;
-
-    private MavenRepoManager mavenRepoManager;
-
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
         if (skip) {
             return;
         }
-        MavenRepositoriesEnricher.enrich(mavenSession, project, repositories);
-        mavenRepoManager = new MavenArtifactRepositoryManager(repoSystem, session, repositories);
-        final Log log = getLog();
-        final Path deploymentContent = getDeploymentContent();
-        // The deployment must exist before we do anything
-        if (Files.notExists(deploymentContent)) {
-            throw new MojoExecutionException(String.format("The deployment '%s' could not be found.", deploymentContent.toAbsolutePath()));
-        }
-        // Validate the environment
-        final Path wildflyPath = provisionIfRequired(deploymentContent.getParent().resolve(provisioningDir));
-        if (!ServerHelper.isValidHomeDirectory(wildflyPath)) {
-            throw new MojoExecutionException(String.format("JBOSS_HOME '%s' is not a valid directory.", wildflyPath));
-        }
-        final StandaloneCommandBuilder commandBuilder = createCommandBuilder(wildflyPath);
-
-        // Print some server information
-        log.info("JAVA_HOME : " + commandBuilder.getJavaHome());
-        log.info("JBOSS_HOME: " + commandBuilder.getWildFlyHome());
-        log.info("JAVA_OPTS : " + Utils.toString(commandBuilder.getJavaOptions(), " "));
         try {
-            if (addUser != null && addUser.hasUsers()) {
-                log.info("Adding users: " + addUser);
-                addUser.addUsers(commandBuilder.getWildFlyHome(), commandBuilder.getJavaHome());
+            final Log log = getLog();
+            final Path deploymentContent = getDeploymentContent();
+            // The deployment must exist before we do anything
+            if (Files.notExists(deploymentContent)) {
+                throw new MojoExecutionException(String.format("The deployment '%s' could not be found.", deploymentContent.toAbsolutePath()));
             }
             // Start the server
             log.info("Server is starting up. Press CTRL + C to stop the server.");
-            Process process = startContainer(commandBuilder);
+            final ServerContext context = startServer(ServerType.STANDALONE);
+            Process process = context.process();
             try (ModelControllerClient client = createClient()) {
                 // Execute commands before the deployment is done
                 final CommandConfiguration cmdConfig = CommandConfiguration.of(this::createClient, this::getClientConfiguration)
                         .addCommands(commands)
                         .addScripts(scripts)
-                        .setJBossHome(commandBuilder.getWildFlyHome())
+                        .setJBossHome(context.jbossHome())
                         .setFork(true)
                         .setStdout("none")
                         .setTimeout(timeout)
@@ -312,7 +163,7 @@ public class RunMojo extends AbstractServerConnection {
                     if (exitCode == 10) {
                         // Ensure the current process is destroyed and restart a new one
                         process.destroy();
-                        process = startContainer(commandBuilder);
+                        process = startServer(ServerType.STANDALONE).process();
                     } else {
                         keepRunning = false;
                     }
@@ -327,87 +178,14 @@ public class RunMojo extends AbstractServerConnection {
         }
     }
 
-    /**
-     * Allows the {@link #javaOpts} to be set as a string. The string is assumed to be space delimited.
-     *
-     * @param value a spaced delimited value of JVM options
-     */
-    @SuppressWarnings("unused")
-    public void setJavaOpts(final String value) {
-        if (value != null) {
-            javaOpts = value.split("\\s+");
-        }
-    }
-
-    private StandaloneCommandBuilder createCommandBuilder(final Path wildflyPath) {
-        final StandaloneCommandBuilder commandBuilder = StandaloneCommandBuilder.of(wildflyPath)
-                .setJavaHome(javaHome)
-                .addModuleDirs(modulesPath.getModulePaths());
-
-        // Set the JVM options
-        if (Utils.isNotNullOrEmpty(javaOpts)) {
-            commandBuilder.setJavaOptions(javaOpts);
-        }
-
-        if (serverConfig != null) {
-            commandBuilder.setServerConfiguration(serverConfig);
-        }
-
-        if (propertiesFile != null) {
-            commandBuilder.setPropertiesFile(propertiesFile);
-        }
-
-        if (serverArgs != null) {
-            commandBuilder.addServerArguments(serverArgs);
-        }
-        return commandBuilder;
-    }
-
-    private Path provisionIfRequired(final Path installDir) throws MojoFailureException {
-        if (jbossHome != null) {
-            //we do not need to download WildFly
-            return Paths.get(jbossHome);
-        }
-        try {
-            if (!Files.exists(installDir)) {
-                getLog().info("Provisioning default server in " + installDir);
-                GalleonUtils.provision(installDir, resolveFeaturePackLocation(), version, mavenRepoManager);
-            }
-            return installDir;
-        } catch (ProvisioningException ex) {
-            throw new MojoFailureException(ex.getLocalizedMessage(), ex);
-        }
+    @Override
+    protected CommandBuilder createCommandBuilder(final Path jbossHome) throws MojoExecutionException {
+        return createStandaloneCommandBuilder(jbossHome, serverConfig);
     }
 
     @Override
     public String goal() {
         return "run";
-    }
-
-    private String resolveFeaturePackLocation() {
-        return featurePackLocation == null ? getDefaultFeaturePackLocation() : featurePackLocation;
-    }
-
-    /**
-     * Returns the default feature pack location if not defined in the configuration.
-     *
-     * @return the default feature pack location
-     */
-    protected String getDefaultFeaturePackLocation() {
-        return "wildfly@maven(org.jboss.universe:community-universe)";
-    }
-
-    private Process startContainer(final CommandBuilder commandBuilder) throws IOException, InterruptedException, TimeoutException {
-        final Launcher launcher = Launcher.of(commandBuilder)
-                .inherit();
-        if (env != null) {
-            launcher.addEnvironmentVariables(env);
-        }
-        final Process process = launcher.launch();
-        try (ModelControllerClient client = createClient()) {
-            ServerHelper.waitForStandalone(process, client, startupTimeout);
-        }
-        return process;
     }
 
     private Path getDeploymentContent() {
